LayoutTests/fast/js/kde/delete-expected.txt

PASS a = 1; delete a; is true
PASS delete nonexistant; is true
PASS delete NaN is false
PASS successfullyParsed is true
